lots of python stuff, pretty sure the simulation is now sort of running in simulation.py

+++ particle.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "particle.h"
 
-struct particle * particle_new(float mass, float radius, struct vector3 *position, struct vector3 *velocity) {
+struct particle * particle_new(unsigned long mass, double radius, struct vector3 *position, struct vector3 *velocity) {
   struct particle *p = (struct particle *)malloc(sizeof(struct particle));
   if (p == NULL) return NULL;
   p->mass = mass;
